global: bump invenio-app-rdm to v11
* add temporal fix to override
  frontpage JS
