--- assets/js/invenio_app_rdm/frontpage/override.js
@@ -1,13 +1,0 @@
-// This file is part of CDS RDM
-// Copyright (C) 2022 CERN.
-//
-// CDS RDM is free software; you can redistribute it and/or modify it
-// under the terms of the MIT License; see LICENSE file for more details.
-
-import { CDSRecordsList } from "../../components/frontpage/overrides/RecordsList";
-import { CDSRecordsResultsListItem } from "../../components/frontpage/overrides/RecordsResultsListItem";
-
-export const overriddenComponents = {
-  "InvenioAppRdm.RecordsList.layout": CDSRecordsList,
-  "InvenioAppRdm.RecordsResultsListItem.layout": CDSRecordsResultsListItem,
-};
