+++ assets/js/invenio_app_rdm/overridableRegistry/mapping.js
@@ -1,0 +1,14 @@
+import { CDSCarouselItem } from "../../components/communities_carousel/overrides/CarouselItem";
+import { CDSCommunitiesCarousel } from "../../components/communities_carousel/overrides/CommunitiesCarousel";
+import { CDSRecordsList } from "../../components/frontpage/overrides/RecordsList";
+import { CDSRecordsResultsListItem } from "../../components/frontpage/overrides/RecordsResultsListItem";
+
+
+export const overriddenComponents = {
+    "InvenioAppRdm.RecordsList.layout": CDSRecordsList,
+    "InvenioAppRdm.RecordsResultsListItem.layout": CDSRecordsResultsListItem,
+    "InvenioCommunities.CommunitiesCarousel.layout": CDSCommunitiesCarousel,
+    "InvenioCommunities.CarouselItem.layout": CDSCarouselItem,
+  };
+
+
