+++ invenio.cfg
@@ -109,6 +109,10 @@ THEME_SITENAME = 'CDS'
 #THEME_FOOTER_TEMPLATE = 'cds_rdm/footer.html'
 THEME_HEADER_TEMPLATE = 'cds_rdm/header.html'
 
+# TEMPORAL FIX - to be removed once the js bundle loading issue on the macros
+# is fixed.
+BASE_TEMPLATE = "cds_rdm/page.html"
+
 
 # Invenio-App-RDM
 # ===============
